Add new demo of "rotating pairs"
Found a new class of periodic solutions: Start with a critical point of the
k vortex Hamiltonian and substitute each of the vortices for a pair of
vortices of the same vorticity, rotated by +/- epsilon. The thus created
vortex pairs will orbit their common centre of vorticity, which will remain
at the same location as the jth vortex in the stationary solution of the
original k vortex problem.